requirements.txt
asgiref==3.4.1
autopep8==1.6.0
betterpath==0.2.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.44.0
django-cors-headers==3.10.0
django-filter==21.1
django-jazzmin==2.4.8
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.2.1
facebook-sdk==3.1.0
Faker==9.9.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
l==0.11.0
Markdown==3.3.6
oauthlib==3.1.1
Pillow==8.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
vcversioner==2.16.0.0
whitenoise==5.3.0
zipp==3.6.0
zope.interface==5.4.0
